requirements_global.txt
# ====================================================================
# == DEPENDENCIAS GLOBALES DEL PROYECTO
# ====================================================================
#
# Este fichero consolida todas las dependencias de los diferentes
# componentes del sistema (Backend y Frontends).
#
# NOTA: Este fichero es solo para fines informativos y de auditoría.
# No debe usarse directamente para instalar dependencias, ya que
# mezcla paquetes de Python (pip) y JavaScript (npm).


# ====================================================================
# == BACKEND (Python - pip)
# ====================================================================
#
# Ubicación: /backend/requirements.txt
#
# ADVERTENCIA: Esta lista parece generada por `pip freeze` en un
# entorno Anaconda y contiene muchas más dependencias de las que el
# proyecto realmente necesita. Se recomienda limpiarla.
#

# -- Dependencias Principales del Proyecto --
Django==5.2.3
djangorestframework==3.16.0
psycopg2-binary==2.9.9
celery==5.5.3
redis==6.2.0
gunicorn==21.2.0
django-cors-headers==4.7.0
drf-yasg==1.21.10

# -- Resto de dependencias del entorno --
aext-assistant @ file:///C:/b/abs_1bfn_ondsd/croot/aext-assistant_1746560154452/work
aext-assistant-server @ file:///C:/b/abs_6f6umrladb/croot/aext-assistant-server_1746552992573/work
# ... (y el resto de las dependencias de anaconda)


# ====================================================================
# == FRONTEND WEB (JavaScript - npm)
# ====================================================================
#
# Ubicación: /frontend-web/package.json
#

# -- Dependencias de Producción (`dependencies`) --
@headlessui/react: ^2.2.4
@hookform/resolvers: ^3.3.4
@tanstack/react-query: ^5.17.9
autoprefixer: ^10.4.16
axios: ^1.4.0
chart.js: ^4.4.7
date-fns: ^4.1.0
framer-motion: ^10.16.4
lucide-react: ^0.429.0
postcss: ^8.4.49
react: ^19.1.0
react-chartjs-2: ^5.2.0
react-datepicker: ^4.10.0
react-dom: ^19.1.0
react-hook-form: ^7.49.2
react-hot-toast: ^2.4.1
react-icons: ^4.10.1
react-router-dom: ^6.15.0
react-toastify: ^9.1.3
recharts: ^2.11.1
tailwindcss: ^3.4.17
yup: ^1.3.3

# -- Dependencias de Desarrollo (`devDependencies`) --
@eslint/js: ^9.30.1
@types/react: ^19.1.8
@types/react-dom: ^19.1.6
@vitejs/plugin-react: ^4.6.0
eslint: ^9.30.1
eslint-plugin-react-hooks: ^5.2.0
eslint-plugin-react-refresh: ^0.4.20
globals: ^16.3.0
vite: ^7.0.3


# ====================================================================
# == FRONTEND DESKTOP (JavaScript - npm)
# ====================================================================
#
# Ubicación: /frontend-desktop/package.json
#

# -- Dependencias de Producción (`dependencies`) --
@electron/remote: ^2.1.3
@emotion/react: ^11.14.0
@emotion/styled: ^11.14.1
@hookform/resolvers: ^5.1.1
@mui/icons-material: ^7.2.0
@mui/material: ^7.2.0
@tanstack/react-query: ^5.83.0
axios: ^1.10.0
chart.js: ^4.5.0
date-fns: ^4.1.0
electron-reload: ^2.0.0-alpha.1
electron-store: ^10.1.0
electron-updater: ^6.6.2
framer-motion: ^12.23.0
lucide-react: ^0.525.0
react: ^19.1.0
react-chartjs-2: ^5.3.0
react-datepicker: ^8.4.0
react-dom: ^19.1.0
react-hook-form: ^7.60.0
react-hot-toast: ^2.5.2
react-icons: ^5.5.0
react-router-dom: ^7.6.3
react-toastify: ^11.0.5
recharts: ^3.1.0
yup: ^1.6.1

# -- Dependencias de Desarrollo (`devDependencies`) --
@eslint/js: ^9.30.1
@types/react: ^19.1.8
@types/react-dom: ^19.1.6
@vitejs/plugin-react: ^4.6.0
autoprefixer: ^10.4.21
concurrently: ^9.2.0
electron: ^37.2.0
electron-builder: ^26.0.12
eslint: ^9.30.1
eslint-plugin-react-hooks: ^5.2.0
eslint-plugin-react-refresh: ^0.4.20
globals: ^16.3.0
postcss: ^8.5.6
tailwindcss: ^3.4.17
vite: ^7.0.3
wait-on: ^8.0.3
